Extract shared test helpers into TestHelpers.swift
Deduplicate makeEditor(), type(), paste(), pressEnter(), pressBackspace(),
and display-inspection helpers that were copied across three test files.
