+++ Tests/MarkdownEditorTests/IntegrationTests.swift
@@ -2,80 +2,6 @@ import Testing
 import AppKit
 @testable import MarkdownEditorCore
 
-// MARK: - Shared Helpers
-
-@MainActor
-private func makeEditor() -> EditorTextView {
-    let textStorage = NSTextStorage()
-    let layoutManager = NSLayoutManager()
-    textStorage.addLayoutManager(layoutManager)
-    let textContainer = NSTextContainer(
-        size: NSSize(width: 500, height: CGFloat.greatestFiniteMagnitude)
-    )
-    textContainer.widthTracksTextView = true
-    layoutManager.addTextContainer(textContainer)
-    return EditorTextView(
-        frame: NSRect(x: 0, y: 0, width: 500, height: 300),
-        textContainer: textContainer
-    )
-}
-
-@MainActor
-private func type(_ text: String, into editor: EditorTextView) {
-    for ch in text {
-        editor.insertText(String(ch), replacementRange: NSRange(location: NSNotFound, length: 0))
-    }
-}
-
-@MainActor
-private func paste(_ text: String, into editor: EditorTextView) {
-    editor.insertText(text, replacementRange: NSRange(location: NSNotFound, length: 0))
-}
-
-@MainActor
-private func pressEnter(in editor: EditorTextView) {
-    editor.insertText("\n", replacementRange: NSRange(location: NSNotFound, length: 0))
-}
-
-/// Returns the text storage string for a specific block's display range.
-@MainActor
-private func displayText(for blockIndex: Int, in editor: EditorTextView) -> String {
-    guard blockIndex < editor.displayRanges.count else { return "" }
-    let range = editor.displayRanges[blockIndex]
-    let ts = editor.textStorage!
-    guard range.upperBound <= ts.length else { return "" }
-    return (ts.string as NSString).substring(with: range)
-}
-
-/// Returns all attributes at a given offset in the text storage.
-@MainActor
-private func attrs(at offset: Int, in editor: EditorTextView) -> [NSAttributedString.Key: Any] {
-    let ts = editor.textStorage!
-    guard offset < ts.length else { return [:] }
-    return ts.attributes(at: offset, effectiveRange: nil)
-}
-
-/// Returns the font at a given offset in the text storage.
-@MainActor
-private func font(at offset: Int, in editor: EditorTextView) -> NSFont? {
-    attrs(at: offset, in: editor)[.font] as? NSFont
-}
-
-/// Returns the foreground color at a given offset in the text storage.
-@MainActor
-private func fgColor(at offset: Int, in editor: EditorTextView) -> NSColor? {
-    attrs(at: offset, in: editor)[.foregroundColor] as? NSColor
-}
-
-/// Switches the active block by placing the cursor at the start of a block
-/// and recomposing.
-@MainActor
-private func activateBlock(_ index: Int, in editor: EditorTextView) {
-    guard index < editor.blocks.count else { return }
-    let rawOffset = editor.blocks[index].range.location
-    editor.recompose(cursorInRaw: rawOffset)
-}
-
 // ============================================================================
 // MARK: - Inline Styling: Active Block
 // ============================================================================
